+++ src/Router.tsx
@@ -1,7 +1,8 @@
 import { BrowserRouter, Route } from 'react-router-dom';
 import { Routes } from 'react-router-dom';
-import MainPage from './pages/main/MainPage';
-import ClientPage from './pages/client/ClientPage';
+import MainPage from './pages/MainPage';
+import ClientPage from './pages/ClientPage';
+import ProductListPage from './pages/ProductListPage';
 
 const Router = () => {
   return (
@@ -9,6 +10,7 @@ const Router = () => {
       <Routes>
         <Route path="/" element={<MainPage />} />
         <Route path="/support" element={<ClientPage />} />
+        <Route path="/products" element={<ProductListPage />} />
       </Routes>
     </BrowserRouter>
   );
