+++ src/pages/ProductListPage.tsx
@@ -1,0 +1,11 @@
+import Navbar from '../components/navbar/Navbar';
+
+const ProductListPage = () => {
+  return (
+    <>
+      <Navbar />
+    </>
+  );
+};
+
+export default ProductListPage;
